feat: Add CV page and model
This commit adds a new route and controller method for the CV page. It also includes a new model, `cv.php`, to represent the CV in the application. The `cv.php` model extends the `Model` class and uses the `HasFactory` trait.

These changes allow users to access the CV page on the website and manage CV data in the database.